+++ docs/test_doc_cookbook_queries.py
@@ -26,10 +26,11 @@ def test_queries():
 def test_how_do_i_override_what_operator_is_used_for_a_query():
     # language=rst
     """
-    .. _Filter.query_operator_to_q_operator:
+    .. _override-operator:
 
     How do I override what operator is used for a query?
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    .. uses Filter.query_operator_to_q_operator
 
     The member `query_operator_to_q_operator` for `Filter` is used to convert from e.g. `:`
     to `icontains`. You can specify another callable here:
@@ -54,10 +55,11 @@ def test_how_do_i_override_what_operator_is_used_for_a_query():
 def test_how_do_i_control_what_q_is_produced(really_big_discography):
     # language=rst
     """
-    .. _Filter.value_to_q:
+    .. _control-q:
 
     How do I control what Q is produced?
     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    .. uses Filter.value_to_q
 
     For more advanced customization you can use `value_to_q`. It is a
     callable that takes `filter, op, value_string_or_f` and returns a
@@ -90,4 +92,31 @@ def test_how_do_i_control_what_q_is_produced(really_big_discography):
 
     # @test
     show_output_collapsed(AlbumTable(), url='/?eighties=0')
+    # @end
+
+
+def test_how_do_I_set_the_name_for_a_filter(big_discography):
+    # language=rst
+    """
+    How do I control the name used in the advanced query?
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    .. uses Filter.query_name
+
+    By default the name of filters are derived from the name you specify or from the model field name
+    For deeply nested names double underscores are replaced with single underscores, and those names
+    can become a bit unwieldy. You can then override this with `query_name`:
+    """
+
+    track_table = Table(
+        auto__model=Track,
+        auto__include=['name', 'album', 'album__artist__name'],
+        columns__album_artist_name__filter=dict(
+            include=True,
+            query_name='artist',
+        ),
+    )
+
+    # @test
+    show_output(track_table, '?-query%2Fquery=artist%3D"black+sabbath"')
     # @end