+++ iommi/docs.py
@@ -1,9 +1,13 @@
 import inspect
 import re
+from collections import defaultdict
 from glob import glob
-from os import stat
+from os.path import join
 from pathlib import Path
-from textwrap import dedent
+from textwrap import (
+    dedent,
+    indent,
+)
 from typing import get_type_hints
 
 from iommi import (
@@ -13,46 +17,81 @@ from iommi import (
 from iommi.base import items
 from iommi.declarative import get_declared
 from iommi.declarative.namespace import (
-    Namespace,
     flatten,
+    Namespace,
 )
 from iommi.refinable import (
     EvaluatedRefinable,
-    SpecialEvaluatedRefinable,
     is_refinable_function,
+    SpecialEvaluatedRefinable,
 )
 from iommi.shortcut import (
     get_shortcuts_by_name,
     is_shortcut,
 )
 from iommi.struct import merged
+from iommi.struct import Struct  # noqa: E402
+from tests.helpers import create_iframe  # noqa: E402
 
 
-def read_cookbook_links():
-    result = []
+def uses_from_cookbooks():
+    import docutils.parsers.rst
+    import docutils.utils
+    import docutils.frontend
+    from docutils.nodes import comment, title, target, section
+    from iommi.declarative.util import strip_prefix
+
+    parser = docutils.parsers.rst.Parser()
+    components = (docutils.parsers.rst.Parser,)
+    settings = docutils.frontend.OptionParser(components=components).get_default_values()
+
+    backrefs = defaultdict(set)
+
     for filename in glob(str(Path(__file__).parent.parent / 'docs' / 'cookbook_*.rst')):
         with open(filename) as f:
-            result.append((parse_cookbook_links(f.readlines()), Path(filename).stem))
-    return result
+            content = f.read()
 
+        document = docutils.utils.new_document(filename, settings=settings)
+        parser.parse(content, document)
 
-def parse_cookbook_links(lines):
-    link_marker = '.. _'
-    anchors = [line.strip()[len(link_marker) :].rstrip(':') for line in lines if line.startswith(link_marker)]
+        targets = [x for x in document.children if isinstance(x, target)]
+        nodes = [x for x in document.children if not isinstance(x, target)]
 
-    # TODO: validate that we only have anchors once
+        target_node = None if not targets else targets[0]
 
-    return {x for x in anchors if not x.endswith('?')}
+        assert len(nodes) == 1, 'Expected a document with only ONE level 1 header'
+        (d,) = nodes
 
+        for node in d.children:
+            if isinstance(node, target):
+                target_node = node
+                continue
+            if not node.attributes['ids']:
+                continue
+            if not node.children:
+                continue
+            if not isinstance(node, section):
+                continue
+            title_node = node.children[0]
+            assert isinstance(title_node, title)
 
-def validate_cookbook_links(cookbook_links):
-    class_by_name = {x.__name__: x for x in get_default_classes()}
+            for c in node.children:
+                x = c.astext()
 
-    for link, name in cookbook_links:
-        if '.' in link:
-            class_name, _, refinable_name = link.partition('.')
-            assert class_name in class_by_name, f'{class_name} was not found in default_classes'
-            getattr(class_by_name[class_name], refinable_name)
+                if isinstance(c, target):
+                    target_node = c
+
+                elif isinstance(c, comment) and x.startswith('uses '):
+                    if x.endswith(':'):
+                        print(f'WARNING: bad `uses`: {x}')
+                    if not target_node:
+                        print(f'WARNING: no target for {title_node.astext()}. It must be ABOVE the title.')
+                        continue
+
+                    backrefs[strip_prefix(x, prefix='uses ')].add((target_node.attributes['ids'][0], title_node.astext()))
+
+    assert backrefs
+    return backrefs
 
 
 def get_default_classes():
@@ -150,7 +189,7 @@ def docstring_param_dict(obj):
     )
 
 
-def indent(levels, s):
+def indent_levels(levels, s):
     return (' ' * levels * 4) + s
 
 
@@ -163,38 +202,27 @@ def get_namespace(c):
     )
 
 
-def get_cookbook_name_by_refinable_name():
-    cookbook_links = read_cookbook_links()
-    validate_cookbook_links(cookbook_links)
-    # TODO: validate that all cookbook links are actually linked somewhere. For example ".. _Column.cells_for_rows" isn't right now.
-    cookbook_name_by_refinable_name = {}
-    for links, name in cookbook_links:
-        for link in links:
-            cookbook_name_by_refinable_name[link] = name
-    return cookbook_name_by_refinable_name
-
-
 def _generate_tests_from_class_docs(classes):
-    cookbook_name_by_refinable_name = get_cookbook_name_by_refinable_name()
+    uses_by_field = uses_from_cookbooks()
 
     for c in classes:
         from io import StringIO
 
         f = StringIO()
-        yield _generate_tests_from_class_doc(f, c, classes, cookbook_name_by_refinable_name)
+        yield _generate_tests_from_class_doc(f, c, classes, uses_by_field)
 
 
-def _generate_tests_from_class_doc(f, c, classes, cookbook_name_by_refinable_name):
+def _generate_tests_from_class_doc(f, c, classes, uses_by_field):
     return (
         inspect.getfile(c),
         'test_doc__api_%s.py' % c.__name__,
-        lambda: _generate_tests_from_class_doc_inner(f, c, classes, cookbook_name_by_refinable_name),
+        lambda: _generate_tests_from_class_doc_inner(f, c, classes, uses_by_field),
     )
 
 
-def _generate_tests_from_class_doc_inner(f, c, classes, cookbook_name_by_refinable_name):
+def _generate_tests_from_class_doc_inner(f, c, classes, uses_by_field):
     def w(levels, s):
-        f.write(indent(levels, s))
+        f.write(indent_levels(levels, s))
         f.write('\n')
 
     def section(level, title, indent=0):
@@ -287,7 +315,7 @@ request = req('get')
             v = get_docs_callable_description(v)
 
             if 'lambda' in v:
-                v = v[v.find('lambda') :]
+                v = v[v.find('lambda'):]
                 v = v.strip().strip(',').replace('\n', ' ').replace('  ', ' ')
         if isinstance(v, Part):
             v = v.bind()
@@ -333,11 +361,13 @@ request = req('get')
             if refinable in defaults:
                 w(0, f'Default: `{default_description(defaults.pop(refinable))}`')
 
-            ref_name = f'{c.__name__}.{refinable}'
-            if ref_name in cookbook_name_by_refinable_name:
+            cookbook_usages = uses_by_field.get(f'{c.__name__}.{refinable}', [])
+            if cookbook_usages:
                 w(0, '')
-                w(0, f'Cookbook: :ref:`{ref_name.lower()}`')
-                w(0, '')
+                w(0, f'Cookbook:')
+                for id_, title in cookbook_usages:
+                    w(1, f':ref:`{id_}`')
+                    w(0, '')
 
         w(0, '')
 
@@ -385,6 +415,14 @@ request = req('get')
                     w(0, f'* `{k}`')
                     w(1, f'* `{v}`')
                 w(0, '')
+
+            cookbook_usages = uses_by_field.get(f'{c.__name__}.{name}', [])
+            if cookbook_usages:
+                w(0, '')
+                w(0, f'Cookbook:')
+                for id_, title in cookbook_usages:
+                    w(1, f':ref:`{id_}`')
+                    w(0, '')
 
     for k, v in get_methods_by_type_by_name(c).items():
         methods = v
@@ -434,3 +472,103 @@ def _print_rst_or_python(doc, w, indent=0):
         w(1, '# language=rst')
         w(1, '"""')
     w(0, '')
+
+
+def write_rst_from_pytest():
+    for source in (Path(__file__).parent.parent / 'docs/').glob('test_*.py'):
+        target = source.parent / f'{source.stem.replace("test_doc__api_", "").replace("test_doc_", "")}.rst'
+
+        with open(source) as source_f:
+            with open(target, 'w') as target_f:
+                rst_from_pytest(source_f, target_f, target)
+
+
+def rst_from_pytest(source_f, target_f, target):
+    blocks = []
+    stack = []
+    state_ = None
+
+    def push(new_state, **kwargs):
+        nonlocal state_
+        nonlocal i
+        assert state_ != 'only test', ('exited @test without @end', source_f.name, i)
+        stack.append(new_state)
+        blocks.append(Struct(state=new_state, lines=[], metadata=kwargs))
+        state_ = new_state
+
+    def pop():
+        nonlocal state_
+        stack.pop()
+        state_ = stack[-1]
+        blocks.append(Struct(state=state_, lines=[], metadata={}))
+
+    def add_line(line):
+        blocks[-1].lines.append(line)
+
+    push('import')
+
+    func_name = None
+    func_count = 0
+
+    for i, line in enumerate(source_f.readlines(), start=1):
+        stripped_line = line.strip()
+        if state_ in ('import', 'py') and line.startswith('def test_'):  # not stripped_line!
+            func_name = line[len('def ') :].partition('(')[0]
+            push('py', func_name=func_name, func_count=0)
+            func_count = 0
+        elif stripped_line.startswith("# language=rst"):
+            push('starting rst')
+        elif stripped_line in ('"""', "'''"):
+            if state_ == 'starting rst':
+                # add_line('')
+                pop()
+                push('rst')
+            elif state_ == 'rst':
+                pop()
+        elif stripped_line.startswith('# @test'):
+            push('only test')
+        elif stripped_line.startswith('# @end'):
+            pop()
+        elif state_ == 'py' and line.startswith(
+            '#'
+        ):  # not stripped_line! skip comments on the global level between functions
+            continue
+        elif state_ == 'py' and line.startswith(
+            '@'
+        ):  # not stripped_line! skip decorators on the global level between functions
+            continue
+        else:
+            if state_ == 'only test':
+                if stripped_line.startswith('show_output(') or stripped_line.startswith('show_output_collapsed('):
+                    name = join(target.stem, func_name)
+                    if func_count:
+                        name += str(func_count)
+                    func_count += 1
+
+                    blocks.append(
+                        Struct(
+                            state='raw',
+                            lines=[create_iframe(name, collapsed=stripped_line.startswith('show_output_collapsed'))],
+                            metadata={},
+                        )
+                    )
+            else:
+                add_line(line)
+
+    for b in blocks:
+        b.text = dedent(''.join(b.lines)).strip()
+        del b['lines']
+
+    blocks = [x for x in blocks if x.text]
+
+    for b in blocks:
+        if b.state == 'rst':
+            target_f.write(b.text)
+        elif b.state == 'py':
+            target_f.write('.. code-block:: python\n\n')
+            target_f.write(indent(b.text, '    '))
+        elif b.state == 'raw':
+            target_f.write('.. raw:: html\n\n')
+            target_f.write(indent(b.text, '    '))
+
+        target_f.write('\n\n')