+++ HISTORY.rst
@@ -1,6 +1,23 @@
 Changelog
 ---------
 
+7.9.1 (2025-01-17)
+~~~~~~~~~~~~~~~~~~
+
+* Fixed crash when specifying `Field.parsed_data` callback and the field was not editable
+
+* Improved error messages for evaluation
+
+7.9.0 (2025-01-14)
+~~~~~~~~~~~~~~~~~~
+
+* Add support for `related_query_name`
+
+* `auto_rowspan` on a generator gives an empty table
+
+* Admin: configurable grouping of models
+
+
 7.8.0 (2024-12-12)
 ~~~~~~~~~~~~~~~~~~